[최종] 개인 캘린더 추가 (#237)
* fix: 스터디 등록 수정

스터디 등록 수정 및 테스트 완료

[skip ci]

* fix: 스터디 수정, 모집상태 수정 기능 수정

스터디 수정 기능 테스트 완료
모집상태 수정 기능 테스트 완료

[skip ci]

* fix: 스터디 리스트 조회 기능 수정

리턴 값에 tagValues List 추가됨.
스터디 리스트 조회 테스트 완료

[skip ci]

* refactor: 시그니처 변경: getLoggedIn(Authentication authentication)

getLoggedIn(Authentication authentication) -> getLoggedIn(String Email)

[skip ci]

* fix: Member Transaction LAZY 문제( CalendarStudygroup )

service 로 email 정보만 전달하도록 수정

[skip ci]

* fix: Member Transaction LAZY 문제( StudygroupPostComment )

service 로 email 정보만 전달하도록 수정

[skip ci]

* fix: Member Transaction LAZY 문제( StudygroupJoin )

service 로 email 정보만 전달하도록 수정

[skip ci]

* feat: CalenderMember 추가

 전체적인 틀을 Studygroup 에서 복사해옴

[skip ci]

* feat: CalenderMember 추가

 전체적인 틀을 Studygroup 에서 복사해옴

[skip ci]

* feat: CalenderMember - create

create 구현 - controller, service, mapper

[skip ci]

* feat: CalenderMember - update

controller, service 구현

[skip ci]

* feat: CalenderMember - get

controller 구현

[skip ci]

* feat: CalenderMember - update, delete

controller, service, repository, util 구현

[skip ci]

* fix: CalenderMember

세부 사항 구현

[skip ci]

* fix: CalenderMember 컬럼 이름 변경

content -> platform , 세부 사항 구현

[skip ci]

* fix: CalenderMember 모호한 엔드포인트 수정

get 동일한거 수정

[skip ci]

* feat: CalenderMember 엔드포인트 수정 및 주석 추가

추가

[skip ci]

* feat: StudygroupPostComment 컬럼명 수정

platform -> content 수정

[skip ci]

* fix: CalendarMapper 매핑 오류 수정

mapper 순서 영향없도록 해주는 의존성 추가
TimeRange 엔티티에 Builder 추가 ( 자동 매핑을 위해서 )

[skip ci]

* fix: 스터디 캘린더 예외 추가 - 조회

일정 조회 시 스터디 식별자에 속하지 않은 경우 예외처리

[skip ci]

* fix: 스터디 삭제 시 캘린더 자동삭제

스터디 삭제 시 캘린더 내용도 함께 삭제

[skip ci]

* feat: 개인 일정 수정 시 예외처리 추가

개인이 스터디 일정은 수정할 수 없음

[skip ci]

* refactor: 검증 로직 utils 로 이동

[skip ci]

* feat: 예외처리 추가

일정의 스터디 ID가 null 일 때

[skip ci]

* feat: CalendarResponseDto 매퍼 추가

[skip ci]

* feat: Calendar 에 color 필드, 컬럼 추가

[skip ci]

* feat: Calendar 에 description 필드, 컬럼 추가

CalendarMemberResponse 에 description 추가

[skip ci]

* fix: Calendar 매핑 안되던 것 수정

create, update 매핑 수정

[skip ci]

---------

+++ server/edusync/src/main/java/com/codestates/edusync/model/study/plancalendar/service/CalendarStudygroupService.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import static com.codestates.edusync.exception.ExceptionCode.YOU_ARE_NOT_STUDYGROUP_LEADER;
+import static com.codestates.edusync.exception.ExceptionCode.*;
 
 @Transactional
 @RequiredArgsConstructor
@@ -31,13 +31,18 @@ public class CalendarStudygroupService implements CalendarStudygroupManager {
     public void createTimeSchedules(Long studygroupId,
                                     List<TimeSchedule> timeSchedules,
                                     String email) {
-        Member loginMember = memberUtils.getLoggedIn(email);
+        if( !studygroupUtils.isMemberLeaderOfStudygroup(email, studygroupId) ) {
+            throw new BusinessLogicException(YOU_ARE_NOT_STUDYGROUP_LEADER);
+        }
+
         Studygroup findStudygroup = studygroupUtils.findVerifyStudygroup(studygroupId);
 
         timeSchedules.forEach(ts -> {
             ts.setStudygroup(findStudygroup);
             ts.setTitle(findStudygroup.getStudyName());
             ts.setPlatform((findStudygroup.getPlatform()));
+            ts.setDescription(findStudygroup.getIntroduction());
+            // todo: color 기본값 적용할 수 있도록
         } );
         calendarRepository.saveAll(timeSchedules);
 
@@ -67,6 +72,8 @@ public class CalendarStudygroupService implements CalendarStudygroupManager {
                             TimeSchedule result = new TimeSchedule();
                             result.setTitle(ts.getTitle());
                             result.setPlatform(ts.getPlatform());
+                            result.setDescription(ts.getDescription());
+                            result.setColor(ts.getColor());
                             result.setTime(ts.getTime());
                             result.setMember(studygroupJoin.getMember());
                             result.setStudygroup(findStudygroup);
@@ -88,6 +95,8 @@ public class CalendarStudygroupService implements CalendarStudygroupManager {
 
         Optional.ofNullable(timeSchedule.getTitle()).ifPresent(findTimeSchedule::setTitle);
         Optional.ofNullable(timeSchedule.getPlatform()).ifPresent(findTimeSchedule::setPlatform);
+        Optional.ofNullable(timeSchedule.getDescription()).ifPresent(findTimeSchedule::setDescription);
+        Optional.ofNullable(timeSchedule.getColor()).ifPresent(findTimeSchedule::setColor);
 
         findTimeSchedule.setTime(
                 new TimeRange(
@@ -127,6 +136,8 @@ public class CalendarStudygroupService implements CalendarStudygroupManager {
         findTimeSchedules.forEach(ts -> {
             Optional.ofNullable(timeSchedule.getTitle()).ifPresent(ts::setTitle);
             Optional.ofNullable(timeSchedule.getPlatform()).ifPresent(ts::setPlatform);
+            Optional.ofNullable(timeSchedule.getDescription()).ifPresent(ts::setDescription);
+            Optional.ofNullable(timeSchedule.getColor()).ifPresent(ts::setColor);
 
             ts.setTime(
                     new TimeRange(
@@ -149,10 +160,19 @@ public class CalendarStudygroupService implements CalendarStudygroupManager {
 
     @Override
     public TimeSchedule getSingleTimeScheduleByTimeScheduleId(Long studygroupId, Long timeScheduleId) {
+        TimeSchedule findTimeSchedule = calendarUtils.findVerifyTimeSchedule(timeScheduleId);
+        if( findTimeSchedule.getStudygroup() == null ) {
+            throw new BusinessLogicException(TIME_SCHEDULE_NOT_LINKED_WITH_STUDYGROUP);
+        }
 
-        return calendarUtils.findVerifyTimeSchedule(timeScheduleId);
+        if( !findTimeSchedule.getStudygroup().getId().equals(studygroupId) ) {
+            throw new BusinessLogicException(TIME_SCHEDULE_NOT_MATCHED_WITH_STUDYGROUP);
+        }
+
+        return findTimeSchedule;
     }
 
+    // todo : 스터디 삭제 시 일정 제거용으로 이걸 호출해야함
     @Override
     public void deleteAllTimeSchedulesByStudygroupId(Long studygroupId,
                                                      String email) {
@@ -171,6 +191,9 @@ public class CalendarStudygroupService implements CalendarStudygroupManager {
                                                    String email) {
         studygroupUtils.isMemberLeaderOfStudygroup(email, studygroupId);
         TimeSchedule findTimeSchedule = calendarUtils.findVerifyTimeSchedule(timeScheduleId);
+        if( !findTimeSchedule.getStudygroup().getId().equals(studygroupId) ) {
+            throw new BusinessLogicException(TIME_SCHEDULE_NOT_MATCHED_WITH_STUDYGROUP);
+        }
 
         calendarRepository.delete(findTimeSchedule);
     }
